refactor(cli): split code command helpers
